refactor: update refactor dependencies
